[capital-for-platforms] Add XS promo tile to homepage if offer is available
